// src/Features/RegisterDoctorForm/Model/consts/consts.ts
// for enums consts wcich compiles in object

export const enum ValidateRegisterDoctorError {
    INCORRECT_FIRST_NAME = 'INCORRECT_FIRST_NAME',
    INCORRECT_LAST_NAME = 'INCORRECT_LAST_NAME',
    INCORRECT_PATRONYMIC = 'INCORRECT_PATRONYMIC',
    INCORRECT_LOGIN = 'INCORRECT_LOGIN ',
    INCORRECT_PASSWORD = 'INCORRECT_PASSWORD',
    INCORRECT_DATE_BIRTHDAY = 'INCORRECT_DATE_BIRTHDAY',
    INCORRECT_ADDRESS = 'INCORRECT_ADDRESS',
    INCORRECT_PHONE_NUMBER = 'INCORRECT_PHONE_NUMBER',
    INCORRECT_EXPERIENCE = 'INCORRECT_EXPERIENCE',
    INCORRECT_CABINET_NUMBER = 'INCORRECT_CABINET_NUMBER',
    NO_DATA = 'NO_DATA',
    SERVER_ERROR = 'SERVER_ERROR',
}